requirements.txt
anthropic>=0.40.0
openai>=1.0.0
psutil>=6.0.0
python-dotenv>=1.0.1
flask>=3.0.0
flask-cors>=4.0.0
reportlab>=4.0.0
